content range and usize

+++ sdk/storage/azure_storage_blob/src/clients/blob_client.rs
@@ -338,7 +338,7 @@ impl<'a> BlobClientDownloadBehavior<'a> {
 #[cfg_attr(not(target_arch = "wasm32"), async_trait)]
 #[cfg_attr(target_arch = "wasm32", async_trait(?Send))]
 impl PartitionedDownloadBehavior for BlobClientDownloadBehavior<'_> {
-    async fn transfer_range(&self, range: Option<Range<u64>>) -> Result<AsyncRawResponse> {
+    async fn transfer_range(&self, range: Option<Range<usize>>) -> Result<AsyncRawResponse> {
         let mut opt = self.options.clone();
         opt.range = range.map(|r| format!("bytes={}-{}", r.start, r.end - 1));
         self.client
